[improvement](chat) Merge HanlpDictMapper and FuzzyNameMapper into KeywordMapper.
