+++ launchers/standalone/src/test/resources/META-INF/spring.factories
@@ -1,7 +1,6 @@
 com.tencent.supersonic.chat.api.component.SchemaMapper=\
     com.tencent.supersonic.chat.mapper.EmbeddingMapper, \
-    com.tencent.supersonic.chat.mapper.HanlpDictMapper, \
-    com.tencent.supersonic.chat.mapper.FuzzyNameMapper, \
+    com.tencent.supersonic.chat.mapper.KeywordMapper, \
     com.tencent.supersonic.chat.mapper.QueryFilterMapper, \
     com.tencent.supersonic.chat.mapper.EntityMapper, \
     com.tencent.supersonic.chat.mapper.ModelClusterMapper
